REFACTOR: TaskSearch component
